+++ frontend/portals/Widget/index.jsx
@@ -2,25 +2,25 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import ProductsSlider from '../../components/ProductSlider';
-import { getRecentlyViewedProductsWithLimit, hasMore } from '../../selectors';
+import { getRecentlyViewedProductIdsWithLimit, hasMore } from '../../selectors';
 
 /**
  * Portal position for Products Slider on PDP.
- * @params {Array} products Products collection.
+ * @params {string[]} productIds Product id collection.
  * @params {bool} showMore Whether to show more button (products.length > totalCount)
  * @returns {JSX}
  */
-const Widget = ({ products, showMore, settings }) => (
+const Widget = ({ productIds, showMore, settings }) => (
   <ProductsSlider
     showMore={showMore}
-    products={products}
+    productIds={productIds}
     headline={settings.headline}
     autoPlay={settings.autoPlay}
   />
 );
 
 Widget.propTypes = {
-  products: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+  productIds: PropTypes.arrayOf(PropTypes.string).isRequired,
   showMore: PropTypes.bool.isRequired,
   settings: PropTypes.shape({}),
 };
@@ -35,7 +35,7 @@ Widget.defaultProps = {
  * @returns {Object}
  */
 const mapStateToProps = state => ({
-  products: getRecentlyViewedProductsWithLimit(state),
+  productIds: getRecentlyViewedProductIdsWithLimit(state),
   showMore: hasMore(state),
 });