+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## [Unreleased]
 
 ## [1.2.2]
-### Fixed
-* Added timeout to allow react-id-swiper to apply styling on PDP
+### Added
+* Use of Theme components for ProductSlider and ProductList
 
 ## [1.2.1]
 ### Fixed